Renaming the jacobiMethodScript.js file to jacobiSolver.js
