+++ src/methods/jacobiSolver.js
